fix: Handle properly batch insertion from hexpm
When we are doing a batch insert of hexpm_snapshot the trigger gets
fired with every single insertion. This turn out being a performance
issue and making the whole operation a lot slower than needed.

Instead we are skipping the trigger when doing a batch insert and
refreshing the materialized view at the end of the process

There are multiple way of disabling the trigger:
One was to alter table and disble the trigger and after the batch
insertion enable the trigger again, this is not useful for our case
because alter table would lock the table and it will affect all the
other users

Other option was to set session replication role, this will disable more
checks than needed

We want to be isolate the code that skip the trigger as much as
possible, so we end up doing a set local that only applies to that
transaction this way we dont need to worry about enabling back because
it will after the transaction

For doing that we need to also update the trigger to skip it based on a
new custom postgres setting